Break logging and TLS config out into separate files
